+++ src/mss/linux/xcb.py
@@ -1,11 +1,13 @@
 from __future__ import annotations
 
-from ctypes import _Pointer, c_int
+import contextlib
+from ctypes import _Pointer, addressof, c_int
+from typing import Literal, overload
 
 from . import xcbgen
 
 # We import these just so they're re-exported to our users.
-# ruff: noqa: F401, TC001
+# ruff: noqa: F401
 from .xcbgen import (
     RANDR_MAJOR_VERSION,
     RANDR_MINOR_VERSION,
@@ -29,12 +31,19 @@ from .xcbgen import (
     ImageOrder,
     Keycode,
     Pixmap,
+    RandrConnection,
     RandrCrtc,
     RandrGetCrtcInfoReply,
+    RandrGetMonitorsReply,
+    RandrGetOutputInfoReply,
+    RandrGetOutputPrimaryReply,
+    RandrGetOutputPropertyReply,
     RandrGetScreenResourcesCurrentReply,
     RandrGetScreenResourcesReply,
     RandrMode,
     RandrModeInfo,
+    RandrMonitorInfo,
+    RandrMonitorInfoIterator,
     RandrOutput,
     RandrQueryVersionReply,
     RandrSetConfig,
@@ -75,6 +84,16 @@ from .xcbgen import (
     randr_get_crtc_info,
     randr_get_crtc_info_outputs,
     randr_get_crtc_info_possible,
+    randr_get_monitors,
+    randr_get_monitors_monitors,
+    randr_get_output_info,
+    randr_get_output_info_clones,
+    randr_get_output_info_crtcs,
+    randr_get_output_info_modes,
+    randr_get_output_info_name,
+    randr_get_output_primary,
+    randr_get_output_property,
+    randr_get_output_property_data,
     randr_get_screen_resources,
     randr_get_screen_resources_crtcs,
     randr_get_screen_resources_current,
@@ -85,6 +104,7 @@ from .xcbgen import (
     randr_get_screen_resources_modes,
     randr_get_screen_resources_names,
     randr_get_screen_resources_outputs,
+    randr_monitor_info_outputs,
     randr_query_version,
     render_pictdepth_visuals,
     render_pictscreen_depths,
@@ -109,7 +129,7 @@ from .xcbgen import (
 )
 
 # These are also here to re-export.
-from .xcbhelpers import LIB, XID, Connection, QueryExtensionReply, XcbExtension, XError
+from .xcbhelpers import LIB, XID, Connection, InternAtomReply, QueryExtensionReply, XcbExtension, XError
 
 XCB_CONN_ERROR = 1
 XCB_CONN_CLOSED_EXT_NOTSUPPORTED = 2
@@ -133,6 +153,147 @@ XCB_CONN_ERRMSG = {
 
 
 #### High-level XCB function wrappers
+
+# XCB_NONE is the universal null resource or null atom parameter value for many core X requests
+XCB_NONE = Atom(0)
+
+# Some atoms are defined by the spec, to avoid apps having to look them up.  It's fine to look them up anyway.
+_PREDEFINED_ATOMS = {
+    "PRIMARY": Atom(1),
+    "SECONDARY": Atom(2),
+    "ARC": Atom(3),
+    "ATOM": Atom(4),
+    "BITMAP": Atom(5),
+    "CARDINAL": Atom(6),
+    "COLORMAP": Atom(7),
+    "CURSOR": Atom(8),
+    "CUT_BUFFER0": Atom(9),
+    "CUT_BUFFER1": Atom(10),
+    "CUT_BUFFER2": Atom(11),
+    "CUT_BUFFER3": Atom(12),
+    "CUT_BUFFER4": Atom(13),
+    "CUT_BUFFER5": Atom(14),
+    "CUT_BUFFER6": Atom(15),
+    "CUT_BUFFER7": Atom(16),
+    "DRAWABLE": Atom(17),
+    "FONT": Atom(18),
+    "INTEGER": Atom(19),
+    "PIXMAP": Atom(20),
+    "POINT": Atom(21),
+    "RECTANGLE": Atom(22),
+    "RESOURCE_MANAGER": Atom(23),
+    "RGB_COLOR_MAP": Atom(24),
+    "RGB_BEST_MAP": Atom(25),
+    "RGB_BLUE_MAP": Atom(26),
+    "RGB_DEFAULT_MAP": Atom(27),
+    "RGB_GRAY_MAP": Atom(28),
+    "RGB_GREEN_MAP": Atom(29),
+    "RGB_RED_MAP": Atom(30),
+    "STRING": Atom(31),
+    "VISUALID": Atom(32),
+    "WINDOW": Atom(33),
+    "WM_COMMAND": Atom(34),
+    "WM_HINTS": Atom(35),
+    "WM_CLIENT_MACHINE": Atom(36),
+    "WM_ICON_NAME": Atom(37),
+    "WM_ICON_SIZE": Atom(38),
+    "WM_NAME": Atom(39),
+    "WM_NORMAL_HINTS": Atom(40),
+    "WM_SIZE_HINTS": Atom(41),
+    "WM_ZOOM_HINTS": Atom(42),
+    "MIN_SPACE": Atom(43),
+    "NORM_SPACE": Atom(44),
+    "MAX_SPACE": Atom(45),
+    "END_SPACE": Atom(46),
+    "SUPERSCRIPT_X": Atom(47),
+    "SUPERSCRIPT_Y": Atom(48),
+    "SUBSCRIPT_X": Atom(49),
+    "SUBSCRIPT_Y": Atom(50),
+    "UNDERLINE_POSITION": Atom(51),
+    "UNDERLINE_THICKNESS": Atom(52),
+    "STRIKEOUT_ASCENT": Atom(53),
+    "STRIKEOUT_DESCENT": Atom(54),
+    "ITALIC_ANGLE": Atom(55),
+    "X_HEIGHT": Atom(56),
+    "QUAD_WIDTH": Atom(57),
+    "WEIGHT": Atom(58),
+    "POINT_SIZE": Atom(59),
+    "RESOLUTION": Atom(60),
+    "COPYRIGHT": Atom(61),
+    "NOTICE": Atom(62),
+    "FONT_NAME": Atom(63),
+    "FAMILY_NAME": Atom(64),
+    "FULL_NAME": Atom(65),
+    "CAP_HEIGHT": Atom(66),
+    "WM_CLASS": Atom(67),
+    "WM_TRANSIENT_FOR": Atom(68),
+}
+
+# The atom cache needs to be per-connection. Rather than keying on a (connection, name) tuple, we use a two-level
+# cache keyed by the integer address of the underlying XCB connection (see ctypes.addressof in intern_atom).
+_ATOM_CACHE: dict[int, dict[str, Atom]] = {}
+
+
+@overload
+def intern_atom(
+    xcb_conn: Connection | _Pointer[Connection],
+    name: str,
+    *,
+    only_if_exists: Literal[False] = False,
+) -> Atom: ...
+@overload
+def intern_atom(
+    xcb_conn: Connection | _Pointer[Connection],
+    name: str,
+    *,
+    only_if_exists: Literal[True] = True,
+) -> Atom | None: ...
+@overload
+def intern_atom(
+    xcb_conn: Connection | _Pointer[Connection],
+    name: str,
+    *,
+    only_if_exists: bool,
+) -> Atom | None: ...
+
+
+def intern_atom(
+    xcb_conn: Connection | _Pointer[Connection],
+    name: str,
+    *,
+    only_if_exists: bool = False,
+) -> Atom | None:
+    if name in _PREDEFINED_ATOMS:
+        return _PREDEFINED_ATOMS[name]
+
+    if isinstance(xcb_conn, _Pointer):
+        # Dereference the pointer before using the cache.
+        xcb_conn = xcb_conn.contents
+    cache_key = addressof(xcb_conn)
+    if cache_key not in _ATOM_CACHE:
+        # This can happen if the connection was closed and its cache cleared, but some code still has a reference to
+        # the connection object.  We could re-create the cache entry, but it's safer to just fail instead of silently
+        # allowing lookups to succeed when they shouldn't.
+        msg = "Connection to X server is closed"
+        raise XError(msg)
+    if name in _ATOM_CACHE[cache_key]:
+        return _ATOM_CACHE[cache_key][name]
+
+    # Atom names are required to be Latin-1, per the X protocol spec, although anything that's not in the XPCS (a
+    # subset of ASCII) is vendor-defined.
+    name_encoded = name.encode("latin_1", errors="strict")
+    cookie = LIB.xcb.xcb_intern_atom(xcb_conn, 1 if only_if_exists else 0, len(name_encoded), name_encoded)
+    atom_as_xid = cookie.reply(xcb_conn).atom
+    if atom_as_xid.value == 0:
+        if not only_if_exists:
+            # This shouldn't be possible.  We at least need to have a path for the type-checker to be happy, though.
+            msg = f"X server failed to intern atom '{name}'"
+            raise XError(msg)
+        # We don't do negative caching, since any app might intern the atom at any time.
+        return None
+    atom = Atom(atom_as_xid.value)
+    _ATOM_CACHE[cache_key][name] = atom
+    return atom
 
 
 def get_extension_data(
@@ -210,13 +371,23 @@ def connect(display: str | bytes | None = None) -> tuple[Connection, int]:
     prefetch_extension_data(conn_p, LIB.shm_id)
     prefetch_extension_data(conn_p, LIB.xfixes_id)
 
+    _ATOM_CACHE[addressof(conn_p.contents)] = {}
+
     return conn_p.contents, pref_screen_num.value
 
 
-def disconnect(conn: Connection) -> None:
-    conn_err = LIB.xcb.xcb_connection_has_error(conn)
+def disconnect(xcb_conn: Connection | _Pointer[Connection]) -> None:
+    if isinstance(xcb_conn, _Pointer):
+        # Dereference the pointer before using the cache.
+        xcb_conn = xcb_conn.contents
+
+    # The cache might already be cleared if the connection had an error, or if disconnect was called multiple times.
+    with contextlib.suppress(KeyError):
+        del _ATOM_CACHE[addressof(xcb_conn)]
+
+    conn_err = LIB.xcb.xcb_connection_has_error(xcb_conn)
     # XCB won't free its connection structures until we disconnect, even in the event of an error.
-    LIB.xcb.xcb_disconnect(conn)
+    LIB.xcb.xcb_disconnect(xcb_conn)
     if conn_err != 0:
         msg = "Connection to X server closed: "
         conn_errmsg = XCB_CONN_ERRMSG.get(conn_err)